x/lottery/client/cli/query_bet_info.go
package cli

import (
	"context"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/naruto0913/lottery/x/lottery/types"
	"github.com/spf13/cobra"
)

func CmdShowBetInfo() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-bet-info",
		Short: "shows betInfo",
		Args:  cobra.NoArgs,
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryGetBetInfoRequest{}

			res, err := queryClient.BetInfo(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
